refactoring

+++ backend/src/main/java/com/savchenko/backend/dto/PageRequestDto.java
@@ -4,6 +4,6 @@ import com.savchenko.backend.dto.filter.FilterDto;
 
 public class PageRequestDto<T extends FilterDto> {
     public T filter;
-    public Long pageSize;
-    public Long pageNumber;
+    public Integer pageSize;
+    public Integer pageNumber;
 }
